requirements.txt
cachetools==5.3.2
diskcache==5.6.3
joblib==1.3.2
matplotlib==3.8.2
numpy==1.24.0
nvitop==1.3.2
opacus==1.4.0
opt_einsum==3.3.0
pandas==2.2.1
pymoo==0.6.1.1
pytz==2023.4
Requests==2.31.0
scikit_learn==1.4.0
scipy==1.12.0
shap==0.43.0
torch_optimizer==0.3.0
torchmetrics==1.3.0.post0
torchsummary==1.5.1
tqdm==4.66.1
wget==3.2
tensorboard==2.15.1
torch==2.1.2
torchinfo==1.8.0
torchvision==0.16.2
nmslib-metabrainz==2.1.2
